Move tightly coupled types to the same file
Go is different than other langauges and allow tightly coupled types to be in the same file. This makes it easier to read code without having to switch back and forth among different files.
